neurovault/requirements.txt
# Core database drivers and ORM
psycopg2-binary
neo4j
SQLAlchemy

# pgvector integration for SQLAlchemy
sqlalchemy-pgvector

# Pydantic for data validation, useful for API layers
pydantic

# NLP and ML libraries for the ingestion pipeline
langchain
sentence-transformers

# API Server
Flask

# Structured Logging
structlog

# Prometheus Metrics
prometheus-client
